Rename swinir -> swinir_model
